fix: moved extension PackageCreationError into Yosemite layer because of swift proposal SE-0364
